Update generator configuration
This change makes a few configuration changes which are meant to make
generation go more smoothly next time.

First, in the `.openapi-generator-ignore` file, I've manually ignored a
few filenames that end up being too long to build into a gem. This is
due to a limitation in Ruby's `gem`. Rather than ignore all files with a
length beyond the max, I've opted to add the filenames manually. This
way, if we do generate something in the future that needs to be added,
it will be necessary to handle those. This prevents us from accidentally
skipping new files.

Second, in `openapi_generator_config.yml`, I've made a few changes:

* I've specified global command options. This hopefully makes running
  the `generate` command smoother in the future, since the command will
  be simpler.
* Under the additional properties (specific to Ruby), I've sorted the
  keys to make it easier to scan the list visually and find specific
  options.
* I've used the approximation operator for the Ruby version so that it's
  easier to rebuild next time, even if the Ruby version has changed.
* I've bumped the overall gem version.
* I've enabled autoloading, which should reduce time to start up and
  memory usage when we use this API client.
* I've added more hands-off commands to the file comment.