Hack spec to force type discrimination
In some of the data models' included objects, the type was specified
using `anyOf`, which allows any type fitting the attributes to be used, even if
there is another, more specific type available. This has been leading to issues
with parsing API results, where included objects end up with generic types that
drop lots of attributes.

To make this change, I ran the following commands:

    sed -i '' 's/anyOf:/oneOf:/' openapi-bundled.yaml
    openapi-generator generate --config openapi_generator_config.yml

This effectively executes a find-and-replace to change all `anyOf` to `oneOf` in
the file, updating all these relationships to depend on a type
discriminator field. In practice, it appears that this field is always present in
my testing, so this should be safe.

I suspect that this is a small bug in the spec, which we mean to ask about.

https://swagger.io/docs/specification/v3_0/data-models/oneof-anyof-allof-not/